--- optimize/__init__.py
@@ -1,5 +1,0 @@
-
-from ObjectiveFunction import ObjectiveFunction
-from Gradient import Gradient
-from Optimizer import Optimizer
-
