--- postprocess/__init__.py
@@ -1,2 +1,0 @@
-
-from Postprocessor import Postprocessor
