--- srom/__init__.py
@@ -1,4 +1,0 @@
-
-from SROM import SROM
-from SROMSurrogate import SROMSurrogate
-from FiniteDifference import FiniteDifference
